unirec: moved headers into include/, fixed distcheck
