+++ unirec/include/Makefile.am
@@ -1,0 +1,2 @@
+SUBDIRS = unirec unirec++
+
