+++ unirec/include/unirec++/Makefile.am
@@ -1,0 +1,17 @@
+unirecppincludedir=$(includedir)/unirec++
+unirecppinclude_HEADERS=bidirectionalInterface.hpp \
+	inputInterface.hpp \
+	ipAddress.hpp \
+	macAddress.hpp \
+	outputInterface.hpp \
+	trapModuleInfo.hpp \
+	unirecArray.hpp \
+	unirecException.hpp \
+	unirec.hpp \
+	unirecRecord.hpp \
+	unirecRecordView.hpp \
+	unirecTypes.hpp \
+	unirecTypeTraits.hpp \
+	urTime.hpp
+
+
