+++ unirec/include/unirec++/trapModuleInfo.hpp
