+++ unirec/ip_prefix_search.c
@@ -47,7 +47,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include "ip_prefix_search.h"
+#include <unirec/ip_prefix_search.h>
 #include "ipps_internal.h"
 
 /**
