+++ unirec/tests/test_time.c
@@ -41,7 +41,7 @@
  *
  */
 
-#include "../ur_time.h"
+#include "../include/unirec/ur_time.h"
 #include <stdio.h>
 #include <stdint.h>
 #include <inttypes.h>
